Add files via upload
I know that these files are quite messy in coding. I apologize for that as I quite a newbie in this field.
The biggest issue that i faced right now is that the input field seems to be endlessly extending its width when the user input a very long data.
I would like that the input field extend vertically (fixed width) when the user input long data . while maintaining the width of the table column and only wrapping the height of the input field.